Add context()
According to Mocha documentation, `context()` is just an alias for
`describe()`. Folks might be expecting it to exist on this API, so I am
adding it.

I considered making it simply alias describe here, but I wanted the
`type` to be "context" so I decided to roll my own copypasta.