Fix an issue causing multiple wraps to exponentially add outer wrappers.
It appears that we were double wrapping MochaWrap instances when calling .use
or .with<Wrapper> multiple times. When calling .use, we were correctly creating
a new instance of MochaWrap with all the correct wrappers, but instead of
returning that new instance, we were appending it (and all its wrappers, which
include all of our current instance's wrappers) to our current instance's
wrappers.

Now we're just extending, and returning the new instance.